+++ installation-bundle/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Contao installation bundle change log
 
+### 4.4.3 (2017-08-16)
+
+ * Warm up the Symfony cache after the database credentials have been set (see #63).
+ * Check if the Contao framework has been initialized when adding the user agent string (see standard-edition#64).
+
 ### 4.4.1 (2017-07-12)
 
  * Correctly set the "overwriteMeta" field during the database update (see contao/core-bundle#888).